fix: Allow `DISABLE_SPEEDY` to be set to `false` (#3289)
* fix: Allow `DISABLE_SPEEDY` to be set to `false`

The old initialization of `DISABLE_SPEEDY` meant that the only way it could be set to a falsey value was if `NODE_ENV` was set to `production` and no other setting was set. Setting `SC_DISABLE_SPEEDY` or the environment variables to false would be overridden by `NODE_ENV`.

The new initialization ensures that if one of these options is set, it’s value is respected.

This change should only affect the computed value of `DISABLE_SPEEDY` if one of the settings has the value `false` or `’false’`. Otherwise it should have the same truthy or falsey value.

Since the expected behavior of setting one of the configuration variables to `false` is to have `DISABLE_SPEEDY` set to false, this should be a fix and not a breaking change.

* test: Add tests for setting `DISABLE_SPEEDY` to `false`

* fix: Ensure `DISABLE_SPEEDY` boolean type

Technically not needed, but makes the tests pass when passing in the string values of `’false’` without needing to modify the compare values. And is a bit cleaner to have it as a boolean type.

* fix: Inline `process.env` references

Per [this comment](https://github.com/styled-components/styled-components/pull/3289#issuecomment-699171576)

* docs: Add changelog entry